organizing your code in sections
A good programming practice is to divide the code into sections, and one of them is the imports section, where all your libraries and modules will be located. Doing this is simple: just leave two lines blank after the last import. In this code, it would be fine to keep all the lines together, but in a big code, organization keeps everything prettier! ✨